WIP: Track search results in Enhanced Ecommerce
Use Google Analytics Enhanced Ecommerce to track the results in site
search. This currently just tracks the list of results, not the one
which a user clicks on.

This change has been made in frontend to help us prototype the Ecommerce
analytics. A full implementation shuold probably live in `static` with
the other analytics code.